# docs/reference/openapi.yaml
openapi: 3.1.0
info:
    title: durable-actors HTTP API
    version: "1"
    description: |
        Manage durable-actors deployments, invoke actors, find actor hosts, and inspect activity.
        When DURABLE_ACTORS_SECRET is set on the server, authenticate backend requests
        with that bearer API key; never expose it to browsers. When unset, no API key is required.
        Project-scoped actor routes work in development and production. Run `durable-actors dev`
        and copy the printed connection settings into your backend's environment. The default
        local project ID is `local`; use it as `project_id` in requests.
        Servers listening beyond localhost warn when the secret is unset but still start.
servers:
    - url: http://127.0.0.1:7100
      description: Local server.
security:
    - ApiKey: []
    - {}
...
paths:
...
    /v1/projects/{project_id}/deployment:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
        put:
            operationId: registerDeployment
            tags: [Deployments]
            summary: Deploy actors
            description: Builds the supplied source image, replaces the current deployment, and restarts its actors while preserving saved state. Local development registers source from the filesystem with imageRef set to local.
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/Deployment" }
                        examples:
                            chat:
                                value:
                                    imageRef: registry.example/customer@sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                                    workingDirectory: /customer
                                    actorEntrypoint: src/actors.ts
                                    secretRefs: []
            responses:
                "200": { $ref: "#/components/responses/Changed" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
        get:
            operationId: getDeployment
            tags: [Deployments]
            summary: Get current deployment
            responses:
                "200":
                    description: Active deployment, without the contract.
                    content:
                        application/json:
                            schema: { $ref: "#/components/schemas/CurrentDeployment" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "404": { $ref: "#/components/responses/NotFound" }
                "500": { $ref: "#/components/responses/Internal" }
        delete:
            operationId: deleteDeployment
            tags: [Deployments]
            summary: Delete deployment
            description: Stops deployed actors and keeps their saved state.
            responses:
                "200": { $ref: "#/components/responses/Changed" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "500": { $ref: "#/components/responses/Internal" }
...
    /v1/projects/{project_id}/sessions:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
        post:
            operationId: issueActorSession
            tags: [Sessions]
            summary: Issue a scoped RPC session
            security: [{ ApiKey: [] }]
            description: Requires the configured administrative secret and an active deployment. The trusted backend must authenticate its caller and check project ACLs first. Sessions authorize all actor types, actor IDs, and published RPC methods in exactly one project. Expiry is capped by the requested absolute deadline, 60 seconds, and the runtime maximum JWT lifetime. Session credentials cannot issue sessions or access administrative APIs or WebSockets.
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: "#/components/schemas/IssueActorSession" }
            responses:
                "200":
                    description: Runtime-signed session and its absolute expiration.
                    headers:
                        Cache-Control: { $ref: "#/components/headers/NoStore" }
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [token, expiresAtMs]
                                properties:
                                    token: { type: string }
                                    expiresAtMs: { type: integer, format: int64 }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "404": { $ref: "#/components/responses/NotFound" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "500": { $ref: "#/components/responses/Internal" }
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/find-actor:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: findActor
            tags: [Actors]
            summary: Find an actor host
            security: [{ ApiKey: [] }, { ActorSession: [] }, {}]
            description: Resolves placement, starts a host if needed, and issues credentials for direct HTTP calls. Requires an active deployment. Accepts the administrative secret or a runtime-issued actor:session credential for this project. Sessions produce actor-specific tickets containing all of the actor's published RPC methods. These tickets cannot outlive the session.
            requestBody: { $ref: "#/components/requestBodies/FindActor" }
            responses:
                "200": { $ref: "#/components/responses/ActorTarget" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "403": { description: Actor has no published RPC methods. }
                "404": { $ref: "#/components/responses/NotFound" }
                "409": { $ref: "#/components/responses/Conflict" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "500": { $ref: "#/components/responses/Internal" }
                "503": { $ref: "#/components/responses/Unavailable" }
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/invoke:
        servers:
            - url: "{actorHost}"
              description: Use the cached route returned by invokeActor or findActor.
              variables:
                  actorHost: { default: "http://127.0.0.1:7101" }
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: invokeActor
            tags: [Actors]
            summary: Resolve and invoke an actor method
            servers:
                - url: "{controlPlane}"
                  description: Resolve and invoke with an API key or project session.
                  variables:
                      controlPlane: { default: "http://127.0.0.1:7100" }
                - url: "{actorHost}"
                  description: Cached invocation through the gateway (or local host) with a signed ticket and ownership epoch.
                  variables:
                      actorHost: { default: "http://127.0.0.1:7101" }
            security: [{ ApiKey: [] }, { ActorSession: [] }, { ActorTicket: [] }, {}]
            description: |
                Without a valid cached target, the SDK sends requestId, method, args and an optional homeRegion to the control plane.
                The control plane resolves placement, forwards one call and returns its outcome together with a scoped host target.
                The SDK caches that target for subsequent invocations and broadcasts. Warm invocations use the cached route with its ticket and ownerEpoch. On Kubernetes this is the shared gateway, which verifies the signed route and forwards to the private host.
                The host validates the actor, host session and ownership epoch before dispatch.
                Socket effects are delivered by the host before a completed response.
                The SDK retries once across both paths only for explicit pre-execution rejection or a refused host connection.
                A control-plane outcome of unauthenticated means the host rejected its ticket before execution; HTTP 401 means the caller credential was rejected.
                Actor failures and ambiguous errors are never automatically replayed. An outcome_unknown error means execution may have occurred.
                requestId identifies an attempt; it does not deduplicate execution.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            oneOf:
                                - { $ref: "#/components/schemas/ActorInvocation" }
                                - { $ref: "#/components/schemas/ControlPlaneInvocation" }
                        examples:
                            resolveAndIncrement:
                                value: { requestId: request-1, method: increment, args: [2], homeRegion: north-america-west }
                            increment:
                                value: { requestId: request-1, ownerEpoch: 3, method: increment, args: [2] }
            responses:
                "200":
                    description: Explicit execution outcome, including actor failures. A completed result includes null for void methods.
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - { $ref: "#/components/schemas/ControlPlaneInvocationReply" }
                                    - { $ref: "#/components/schemas/ActorInvocationReply" }
                            examples:
                                resolved:
                                    value:
                                        target: { homeRegion: north-america-west, route: "https://actors.example.com", token: host-ticket, ownerEpoch: 3, expiresAtMs: 1900000000000 }
                                        outcome: { type: completed, result: 7 }
                                completed: { value: { type: completed, result: 7 } }
                                failed: { value: { type: failed, code: actor_error, message: Method failed } }
                                notExecuted: { value: { type: not_executed, reason: host_unavailable } }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { description: "API key, project session or host ticket rejected before dispatch." }
                "403": { description: "Host ticket belongs to another actor, host session or ownership epoch." }
                "404": { $ref: "#/components/responses/NotFound" }
                "409": { $ref: "#/components/responses/Conflict" }
                "502": { description: The invocation outcome is unknown. Do not replay automatically. }
                "503": { $ref: "#/components/responses/Unavailable" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "415": { description: Expected application/json. }
                "422": { description: Invalid invocation document. }
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/socket-effects:
        servers:
            - url: "{actorHost}"
              description: Use the route returned by findActor.
              variables:
                  actorHost: { default: "http://127.0.0.1:7101" }
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: publishActorSocketEffects
            tags: [WebSockets]
            summary: Deliver socket effects to the owning host
            security: [{ ActorTicket: [] }]
            description: Used by backend broadcasts. Actor method effects are already delivered by invokeActor. Do not retry an uncertain delivery.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            additionalProperties: false
                            required: [ownerEpoch, effects]
                            properties:
                                ownerEpoch: { type: integer, minimum: 1, maximum: 9007199254740991 }
                                effects:
                                    type: array
                                    items: { $ref: "#/components/schemas/ActorSocketEffect" }
                        examples:
                            broadcast:
                                value:
                                    ownerEpoch: 3
                                    effects:
                                        - type: broadcast
                                          message: { type: text, data: '"hello"' }
                                          except_connection_ids: []
                                          tags: []
            responses:
                "204": { description: Effects delivered. }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { description: Ticket rejected. }
                "403": { description: "Ticket belongs to another actor, host session or ownership epoch." }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "415": { description: Expected application/json. }
                "422": { description: Invalid effects document. }
                "503": { description: Delivery unavailable; some effects may already have been applied. }
    /v1/projects/{project_id}/actors/{actor_name}/{actor_id}/find-websocket:
        parameters:
            - { $ref: "#/components/parameters/ProjectId" }
            - { $ref: "#/components/parameters/ActorName" }
            - { $ref: "#/components/parameters/ActorId" }
        post:
            operationId: findWebSocket
            tags: [Actors]
            summary: Find an authorized WebSocket URL
            description: Issues a fresh connection URL. Authorize the application user before requesting it. Requires an active deployment.
            requestBody: { $ref: "#/components/requestBodies/FindWebSocket" }
            responses:
                "200": { $ref: "#/components/responses/WebSocketGrant" }
                "400": { $ref: "#/components/responses/InvalidRequest" }
                "401": { $ref: "#/components/responses/Unauthenticated" }
                "409": { $ref: "#/components/responses/Conflict" }
                "413": { $ref: "#/components/responses/PayloadTooLarge" }
                "500": { $ref: "#/components/responses/Internal" }
                "503": { $ref: "#/components/responses/Unavailable" }
...
components:
    securitySchemes:
        ApiKey:
            type: http
            scheme: bearer
            description: Required when DURABLE_ACTORS_SECRET is set on the server. Never send it to browsers.
        ActorTicket:
            type: http
            scheme: bearer
            description: Short-lived actor capability returned by findActor. Never use the application API key at the actor host.
        ActorSession:
            type: http
            scheme: bearer
            description: Runtime-issued project-scoped session, accepted only for actor RPC discovery.
        SocketTicket:
            type: apiKey
            in: query
            name: key
            description: Included in websocketUrl. Keep the URL private.
...
    parameters:
...
        ProjectId:
            name: project_id
            in: path
            required: true
            schema: { type: string, pattern: "^[A-Za-z0-9._-]{1,64}$" }
...
    responses:
        Changed:
            description: Whether the deployment changed.
            content:
                application/json:
                    schema:
                        type: object
                        required: [changed]
                        properties:
                            changed: { type: boolean }
...
        InvalidRequest:
            description: Invalid request (invalid_request).
            content:
                application/json:
                    schema: { $ref: "#/components/schemas/Error" }
...